Add ability to skip tests: skip() and skipIf()
This commit introduces two new functions skip(fmt, ...) and
skipIf(<condition>, fmt, ...) that allow "skipping" (or more
precisely: prematurely ending) tests.

They are meant for tests that might not be relevant under all
conditions (or expected to be unsuccessful for a variety of
reasons). A skipped test results in the new "SKIPPED" node
status, but otherwise is considered 'okay', i.e. passing.